+++ Sources/SwiftKit/_Deprecated/StoreContext+Products.swift
@@ -8,19 +8,14 @@
 
 import StoreKit
 
+@available(*, deprecated, message: "StoreContext has been moved to StoreKitPlus - https://github.com/danielsaidi/StoreKitPlus")
 @available(iOS 15.0, macOS 12.0, watchOS 8.0, tvOS 15.0, *)
 public extension StoreContext {
     
-    /**
-     Whether or not a certain product has an active purchase.
-     */
     func isProductPurchased(id: String) -> Bool {
         purchasedProductIds.contains(id)
     }
     
-    /**
-     Whether or not a certain product has an active purchase.
-     */
     func isProductPurchased(_ product: Product) -> Bool {
         isProductPurchased(id: product.id)
     }