+++ Release Notes.md
@@ -11,7 +11,8 @@ This version adjusts the library for Xcode 14 and deprecates some things.
 
 ### 🗑 Deprecations
 
-* The `StoreKit` namespace has been deprecated and moved to a new library: https://github.com/danielsaidi/StoreKitPlus
+* The `IoC` namespace has been deprecated and will be removed in the next major version.
+* The `StoreKit` namespace has been deprecated and moved to https://github.com/danielsaidi/StoreKitPlus