+++ Release Notes.md
@@ -14,6 +14,7 @@ This version adjusts the library for Xcode 14 and deprecates some things.
 * The `IoC` namespace has been deprecated and will be removed in the next major version.
 * The `Messaging` namespace has been deprecated and will be removed in the next major version.
 * The `StoreKit` namespace has been deprecated and moved to https://github.com/danielsaidi/StoreKitPlus
+* `String+Slugified` has been deprecated and moved to https://github.com/danielsaidi/TagKit